refactor: Stepper common->view 폴더 이동
